+++ workspaces/adr/plugins/adr-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-adr-common
 
+## 0.18.0
+
+### Minor Changes
+
+- da3bcc2: Backstage version bump to v1.49.2
+
 ## 0.17.1
 
 ### Patch Changes